refactor: WIP commit, few files left to migrate and still to sort the templating
