+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.CQRS/xxENSONOxx.xxSTACKSxx.CQRS.csproj
@@ -7,7 +7,8 @@
     <ItemGroup>
         <PackageReference Include="System.ComponentModel.Annotations" Version="5.0.0" />
         <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
-        <PackageReference Include="AWSSDK.SimpleNotificationService" Version="3.7.400.5" Condition="'$(EventPublisherAwsSns)' == 'true'" />
+        <PackageReference Include="AWSSDK.SimpleNotificationService" Version="3.7.400.5" />
+<!--        <PackageReference Include="AWSSDK.SimpleNotificationService" Version="3.7.400.5" Condition="'$(EventPublisherAwsSns)' == 'true'" />-->
     </ItemGroup>
 
     <ItemGroup>
