+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure.UnitTests/SnsEventPublisherTests.cs
