+++ src/cqrs/src/api/xxENSONOxx.xxSTACKSxx.Infrastructure/Serialization/IMessageReader.cs
@@ -1,0 +1,8 @@
+using Azure.Messaging.EventHubs;
+
+namespace xxENSONOxx.xxSTACKSxx.Infrastructure.Serialization;
+
+public interface IMessageReader
+{
+    T Read<T>(EventData eventData);
+}
