+++ src/func-aeh-listener/src/functions/xxENSONOxx.xxSTACKSxx.Listener/Serialization/IMessageReader.cs
@@ -1,0 +1,8 @@
+using Azure.Messaging.EventHubs;
+
+namespace xxENSONOxx.xxSTACKSxx.Listener.Serialization;
+
+public interface IMessageReader
+{
+    T Read<T>(EventData eventData);
+}
