--- src/shared/xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.EventHub/Serializers/IMessageReader.cs
@@ -1,9 +1,0 @@
-﻿using Azure.Messaging.EventHubs;
-
-namespace xxENSONOxx.xxSTACKSxx.Shared.Messaging.Azure.EventHub.Serializers
-{
-    public interface IMessageReader
-    {
-        T Read<T>(EventData eventData);
-    }
-}
